### 2022-03-19 #4256 @editorialbot
[PRE REVIEW]: Spiner: Performance Portable Routines for Generic, Tabulated, Multi-Dimensional Data
**Submitting author:** @Yurlungur (<a href="http://orcid.org/0000-0001-6432-7860">Jonah Miller</a>)
**Repository:** https://github.com/lanl/spiner
**Branch with paper.md** (empty if default branch): joss-paper
**Version:** 1.3.0
