### 2022-03-20 #4256 @xuanxu
@kthyng The [joss-paper branch](https://github.com/lanl/spiner/tree/joss-paper) does not include the code, only the paper files.
