### 2022-03-21 #4256 @danielskatz
@xuanxu (and @arfon) - My thought here is to change the branch to main, run check repository, then change the branch back.  This seems painful and maybe error-prone. How do you think we should handle cases like this, where the author provides a custom branch with the paper, but we want to check the content of the code that is not in that branch?

I'll do my painful thing for now, so we can move this onward, while we talk about a general solution, perhaps outside this issue
