### 2022-03-21 #4256 @danielskatz
@Yurlungur - Due to the relatively small amount of code, the editors will now discuss if it meets the [substantial scholarly effort](https://joss.readthedocs.io/en/latest/submitting.html#substantial-scholarly-effort) criterion for review by JOSS.  You should hear back in a week or so.

In addition, you could work on the possibly missing DOIs that editorialbot suggests, but note that some may be incorrect. Please feel free to make changes to your .bib file, then use the command `@editorialbot check references` to check again, and the command `@editorialbot generate pdf` when the references are right to make a new PDF. editorialbot commands need to be the first entry in a new comment.
