+++ paper.bib
@@ -51,6 +51,7 @@ year = "1992"}
   number={11},
   pages={114103},
   url={http://stacks.iop.org/0264-9381/27/i=11/a=114103},
+  doi={10.1088/0264-9381/27/11/114103},
   year={2010},
 }
 
