### 2022-03-22 #4256 @Yurlungur
@danielskatz sorry for the delay---it's been a busy last couple of days. Here's why I believe the code satisfies the substantial scholarly effort criterion:

The JOSS criteria for substantial scholarly effort is that "Your software should be a significant contribution to the available open source software that either enables some new research challenges to be addressed or makes addressing research challenges significantly better (e.g., faster, easier, simpler)."

Spiner was motivated by an emerging need that we encountered while writing a number of simulation codes, both radiation transport and fluid dynamics, for GPUs: In particular, the performant interpolation of multi-dimensional tabulated data in a portable way. This is a near universal need when dealing with realistic microphysics such as neutrino emissivities and opacities or the equations of state of real materials, such as iron, water, or steel. Thus, a tool in this space has the opportunity to be very impactful by making research challenges, i.e., simulations necessitating the computational power of a GPU node or cluster, significantly easier to overcome. We have already begun leveraging Spiner in two of our own projects, [singularity-eos](https://github.com/lanl/singularity-eos) and [singularity-opac](https://github.com/lanl/singularity-opac), which in turn are leveraged in two continuum dynamics codes and counting. See, the open-source [Phoebus](https://github.com/lanl/phoebus) code for an example. Given this promising start, we believe Spiner can achieve adoption in relevant communities.

A key requirement for broad adoption in this community is that the tool both be lightweight and flexible. Codes use different performance portability strategies and data access patterns, which Spiner needs to support. On the other hand, developers, of both production engineering codes and astrophysics simulation codes (which are the two communities we are most familiar with), are typically leery of taking on large dependencies they cannot understand. We believe Spiner strikes an appropriate balance.

This is why we chose a header-only library and worked hard to avoid software bloat. Spiner is approximately 1600 lines of C++ code including tests, and leverages about 700 lines of additional C++ code in the [ports-of-call](https://github.com/lanl/ports-of-call) library, which we separated out of Spiner for code re-use. The rule of thumb for lines of code in JOSS is roughly 1000, which Spiner exceeds by a fair margin. Nevertheless, it's true that this is a small number of lines. This is intentional, and Spiner underwent many iterations before these particular lines were selected to be open-sourced.

Similarly, this is why we devoted significant time to documentation and to designing a flexible build system with `cmake` and supporting robust package management using `spack`. We wanted downstream users to be able to easily include Spiner in the ways that make sense for their projects.

Finally we note that the commit history on github does not fully reflect Spiner's development. Spiner was developed at Los Alamos National Lab, which has some institutional constraints on open software development, and we were not able to start development out in the open. Rather, Spiner was initially developed internally, starting in 2019, before we received permission to open-source it. This is why, for example, there are several authors on the paper who do not have commits visible in git history.
