### 2022-03-22 #4256 @danielskatz
> I submitted just for Spiner---it didn't occur to me to bundle the two codes for the purpose of the submission. If that's more appropriate, I could revise to do so. They're separate git repositories.

I'm not sure that it's more appropriate, but it would clearly be a sufficiently substantial scholarly effort.  If you want to do that, we can review multiple repositories at once, as long as it's clear in the submission.  The way to do that would be to tell me you want to withdraw this submission, and then make a new one.

On the other hand, I think you've convinced me that this (Spiner) is in sufficient for us to review, but I want to give others a chance to weigh in too, which likely would happen early next week, so you could just wait and see.
