### 2022-03-31 bbe5c01
Sync with spack upstream (#27)
* Sync spackages with spack upstream versions

* update doc

Co-authored-by: Richard Berger <rberger@lanl.gov>
Co-authored-by: Jonah Maxwell Miller <jonahm@lanl.gov>
