### 2022-04-06 #4256 @Yurlungur
Hi @danielskatz I just wanted to reach out to see if a decision has been made regarding substantial scholarly effort for spiner? Thanks!
