### 2022-04-06 #4256 @Kevin-Mattheus-Moerman
@Yurlungur I have just removed the `query scope` label since it was determined this work may be in scope for JOSS. We will now proceed to find a handling editor for this work. You may see me add the `waitlisted` label. This may be needed if none of he editors in this domain are currently available.
