+++ .gitlab-ci.yml
@@ -153,7 +153,7 @@ before_script:
             -DSPINER_USE_CUDA=$([[ ${CI_JOB_NAME} =~ "power9" ]] && echo ON || echo OFF) \
             -DCMAKE_CXX_COMPILER=$([[ ${CI_JOB_NAME} =~ "power9" ]] && echo nvcc_wrapper || g++) \
             ..
-    - make -j
+    - VERBOSE=1 make -j
     - make install
     - make test
     - make convergence
