+++ .gitlab-ci.yml
@@ -148,6 +148,7 @@ before_script:
     - |
       cmake -DBUILD_TESTING=ON \
             -DCMAKE_INSTALL_PREFIX=${CI_PROJECT_DIR}/install \
+            -DSPINER_FORCE_INTERNAL_PORTS=ON \
             -DSPINER_USE_HDF=ON \
             -DSPINER_USE_KOKKOS=$([[ ${CI_JOB_NAME} =~ "power9" ]] && echo ON || echo OFF) \
             -DSPINER_USE_CUDA=$([[ ${CI_JOB_NAME} =~ "power9" ]] && echo ON || echo OFF) \
