+++ CMakeLists.txt
@@ -84,8 +84,6 @@ endif()
 
 if (SPINER_USE_KOKKOS)
   set(PORTABILITY_STRATEGY_KOKKOS)
-  # target_compile_definitions(${SPLIB} INTERFACE
-  #   PORTABILITY_STRATEGY_KOKKOS)
   # Import Kokkos if not already available as a build target
   if (NOT TARGET Kokkos::kokkos)
     if (SPINER_USE_KOKKOS_SRC)
