+++ CMakeLists.txt
@@ -83,7 +83,7 @@ if(SPINER_USE_HDF)
 endif()
 
 if (SPINER_USE_KOKKOS)
-  set(PORTABILITY_STRATEGY "Kokkos" STRING CACHE "Portability strategy" FORCE)
+  set(PORTABILITY_STRATEGY "Kokkos" CACHE STRING "Portability strategy" FORCE)
   # Import Kokkos if not already available as a build target
   if (NOT TARGET Kokkos::kokkos)
     if (SPINER_USE_KOKKOS_SRC)
