### 2022-05-19 #4367 @jzrake
@Yurlungur -- I hope I'm doing this right. Here are two bits of feedback based on my first impression from the PDF writeup.

1. GPU hardware is designed to do interpolation (1d/2d/3d) in hardware as a primitive operation, long before GPGPU was a thing. I think the writeup should state why a software library to do interpolation on the GPU is needed, or is somehow better than what GPUs do in the hardware pipeline.
2. The Statement of Need makes it seem like the primary purpose of the library is for reading opacity tables in hydro codes, which is a relatively specialized application. Is there something about this library which is aimed specifically at such applications, or is the library also intended to offer better (faster?) interpolation for other applications, e.g. image resampling? If it's the specific application, then the Statement of Need should mention (if it's true) that these hydro codes are currently limited either by the accuracy or performance of lookups in the opacity and/or EOS tables.
