### 2022-05-24 #4367 @Yurlungur
@jzrake thanks for the comments. I just updated the manuscript based on your feedback. The relevant commit is [here](https://github.com/lanl/spiner/commit/ba0fb286e1d2c75647ceaf12bbb003cd72ffe088). Here's a formal little writeup to explain the answers to your questions:

1. It's true that hardware interpolation is required for graphics applications, and is thus a feature for GPUs. Indeed, one of our team members applied texture interpolation [as early as 2007](https://www.lanl.gov/orgs/adtsc/publications/nw_highlights_2007/ch1/ch1sec1/1_1_3sriram.pdf). However, we felt a software layer was required for our (and other) scientific applications for several reasons. Texture interpolation, at least on NVIDIA devices, is only single-precision, with interpolation coefficients stored at half-precission. This is often insufficient for scientific applications. Texture interpolation is also rather constrained in application, to only a few stencil patterns and to uniform data only. While `Spiner` currently limits itself to linear interpolation on uniform data, we wanted to leave the door open to other algorithms. Texture interpolation also does not support multi-dimensional mixed indexing/interpoaltion operations where, say, three indices of a four-dimensional array are interpolated and one is merely indexed into. Texture interpolation also, by design operates performantly on vectors of data only, rather than on a single element. While obviously GPUs are vector machines, downstream applications may want to build more complicated operations on scalar interpoaltion primitives. For example, equation of state lookups often involve a root find on interpolated data, which is easier to reason about in scalar form. Finally, the intent of `Spiner` is that the same code base can be used both on CPU and GPU, and on whatever comes next. In other words, that the code be portable. This necessitates a software layer of some kind. That said, a specialization of `Spiner` that uses hardware intrinsics when appropriate would be an interesting topic of future work. We have added comments to the manuscript emphasizing these points.

2. We wrote `Spiner` out of a specific need for such a capability for equation of state and opacity data for continuum dynamics codes, and thus that has been our focus. To our knowledge there is no such standalone capability in the literature, although individual codes have certainly come up with their own internal solutions. Thus we believe our work fills a gap by being a "plug and play" capability for these codes that does not sacrifice performance. Continuum dynamics codes of this kind are broadly applicable to a large number of precision scientific applications, including but not limited to astrophysics, geophysics, climate modeling, and simulations of interest to national defense. Together these applications use up a very large number of supercomputer cycles available. That said, interpolation is of course a very broad topic and as ou point out `Spiner` likely has applications beyond hydro codes. However, we haven't thought very deeply about other pplications. We have added a sentence in the statement of need emphasizing that there is no other performance-portable standalone capability that we know of. We also added a comment about other applications.
