### 2022-05-26 #4367 @jzrake
@Yurlungur thank you for the thoughtful reply. I hope this content can be included in the PDF writeup, as I think it significantly clarifies the motivation and objectives of the project.

@dfm is the PDF writeup intended to be very concise, i.e. can it be as detailed as the reply above?
