### 2022-05-26 #4367 @dfm
@Yurlungur, @jzrake — Re: manuscript scope. Our goal here is for the manuscript to be brief with the documentation page being the primary standalone source of information, so it's generally a good idea to add some words to the documentation whenever you're extending the manuscript. I prefer the manuscript to not have any unique material (everything should be in the docs one way or another), but that's just my preference and not a requirement!
