### 2022-05-27 570722f
update docs to include information in joss paper
