### 2022-06-09 lanl/spiner#34 @lgarrison
[JOSS REVIEW] column-major or row-major?
The [docs](https://lanl.github.io/spiner/main/src/databox.html#creating-a-databox) say that `DataBox` is ordered column-major. But then the example given is the following:
> ```c++
> Spiner::DataBox db(nx3, nx2, nx1);
> ```
> x3 is the slowest moving index and x1 is the fastest

To me, that's row-major indexing. For example, a NumPy array in C-contiguous order (i.e. row major) is indexed as `arr[i, j, k]`, where `i` is the slowest index and `k` is the fastest. This looks like the same ordering of slow-to-fast as `DataBox`, so doesn't that make it row-major?

x-ref https://github.com/openjournals/joss-reviews/issues/4367
