+++ .github/workflows/docs.yml
@@ -11,7 +11,6 @@ on:
 
 jobs:
     docs:
-      if: github.repository_owner == 'lanl'
       name: build and deploy docs
       runs-on: ubuntu-latest
 
@@ -32,6 +31,7 @@ jobs:
             cd doc/sphinx
             make multiversion
         - name: Deploy
+          if: ${{ github.repository == 'lanl/spiner' }}
           uses: peaceiris/actions-gh-pages@v3.7.3
           with:
             github_token: ${{ secrets.GITHUB_TOKEN }}
