### 2022-06-15 f35715f
print some debug info
