+++ .github/workflows/docs.yml
@@ -28,6 +28,7 @@ jobs:
             pip install sphinx-multiversion
         - name: build docs
           run: |
+            echo "Repo = ${GITHUB_REPOSITORY}"
             cd doc/sphinx
             make multiversion
         - name: Deploy
