### 2022-06-15 daa8e43
try using github event trigger
