+++ .github/workflows/docs.yml
@@ -32,7 +32,7 @@ jobs:
             cd doc/sphinx
             make multiversion
         - name: Deploy
-          if: ${{ github.repository == 'lanl/spiner' }}
+          if: ${{ if: github.event.pull_request.head.repo.full_name == github.repository }}
           uses: peaceiris/actions-gh-pages@v3.7.3
           with:
             github_token: ${{ secrets.GITHUB_TOKEN }}
