### 2022-06-15 lanl/spiner#38 @lgarrison
`benchmark.bin` fails silently for large grid size
Per the title, the benchmark seems to complete instantly if the fine grid gets too big:
```console
lgarrison@workergpu142:~/spiner/bin/test$ ./benchmark.bin 64 6000
# ncoarse = 64
# nfine time/point (ns) L2 error
6000    2.03796296296296e-07    0.00000000000000e+00
lgarrison@workergpu142:~/spiner/bin/test$ echo $?
0
```

x-ref: https://github.com/openjournals/joss-reviews/issues/4367, but this is low-priority and doesn't need to be resolved as part of the review.
