### 2022-06-15 lanl/spiner#38 @Yurlungur
Interesting. I'll take a look at this. In what mode did you run this? On GPU? My guess is that the GPU is running out of memory when the benchmark attempts to allocate the arrays. But the fact there's no runtime error is disturbing.
