### 2022-06-15 lanl/spiner#38 @lgarrison
Yes, Kokkos CUDA.

On Wed, Jun 15, 2022 at 2:43 PM Jonah Miller ***@***.***>
wrote:

> Interesting. I'll take a look at this. In what mode did you run this? On
> GPU? My guess is that the GPU is running out of memory when the benchmark
> attempts to allocate the arrays. But the fact there's no runtime error is
> disturbing.
>
