### 2022-06-16 #4367 @Yurlungur
Thanks for quickly digging in to all of that, @lgarrison ! I just updated the paper to address your comments. Here's a more detailed discussion:

Regarding the performance results on CPU Vs. GPU and context for the performance of Spiner: We believe that the phrase "speedup" is misleading, as the goal here is not to speed up an EOS calculation in a CPU code by offloading it to GPU, although that's a benefit of the library. Rather the goal is to provide an interpolation capability that runs natively on CPU and GPU so that a GPU-native code (for example) can simply call the interpolation routines from device kernels. We have thus removed the discussion of a "speedup" from the performance section and instead replaced it with, as you suggested, a
comparison to peak performance. On GPUs, memory bandwidth is often the limiting factor, so we compare to both total peak FLOPS and peak memory movement including both reads and writes. By the former measure, we're about 15% of peak. In the latter case, we achieve about 200 GB/s both read and write. The maximum bandwidth of the device is 900 GB/s.

That said, we thank you for the suggestion that we add some context about performance in this context. We now note in the "state of the field" section that in radiation hydrodynamics calculations, root finding and microphysics subroutines, with interpolation as the innermost operation can take up to 10% of the runtime. For pure hydrodynamics calculations without expensive radiation transport, the cost of evaluating tabulated equations of state can be significantly larger. Without care, it can be as large as half the runtime.

The performance test was indeed done at double precision.
