### 2022-06-29 f5934bc
Small edits to JOSS paper
