+++ paper.bib
@@ -1,7 +1,7 @@
 @ARTICLE{fornax,
        author = {{Skinner}, M. Aaron and {Dolence}, Joshua C. and {Burrows}, Adam and {Radice}, David and {Vartanyan}, David},
         title = "{FORNAX: A Flexible Code for Multiphysics Astrophysical Simulations}",
-      journal = {\apjs},
+      journal = {The Astrophysical Journal Supplement Series},
      keywords = {methods: numerical, Astrophysics - Instrumentation and Methods for Astrophysics, Astrophysics - High Energy Astrophysical Phenomena, Astrophysics - Solar and Stellar Astrophysics, Physics - Computational Physics},
          year = 2019,
         month = mar,
@@ -20,7 +20,7 @@ archivePrefix = {arXiv},
 @ARTICLE{nubhlight,
        author = {{Miller}, Jonah M. and {Ryan}, Ben. R. and {Dolence}, Joshua C.},
         title = "{{\ensuremath{\nu}} bhlight: Radiation GRMHD for Neutrino-driven Accretion Flows}",
-      journal = {\apjs},
+      journal = {The Astrophysical Journal Supplement Series},
      keywords = {accretion, accretion disks, black hole physics, magnetohydrodynamics: MHD, methods: numerical, neutrinos, radiative transfer, Astrophysics - Instrumentation and Methods for Astrophysics, General Relativity and Quantum Cosmology, Physics - Computational Physics},
          year = 2019,
         month = apr,
@@ -118,7 +118,7 @@ url = {https://github.com/catchorg/Catch2},
 @ARTICLE{SullivanWeak,
        author = {{Sullivan}, Chris and {O'Connor}, Evan and {Zegers}, Remco G.~T. and {Grubb}, Thomas and {Austin}, Sam M.},
         title = "{The Sensitivity of Core-collapse Supernovae to Nuclear Electron Capture}",
-      journal = {\apj},
+      journal = {The Astrophysical Journal},
      keywords = {neutrinos, nuclear reactions, nucleosynthesis, abundances, equation of state, supernovae: general, Astrophysics - High Energy Astrophysical Phenomena, Nuclear Experiment, Nuclear Theory},
          year = 2016,
         month = jan,
