### 2022-06-29 #4367 @dfm
@jzrake, @lgarrison — Thanks for your thorough and constructive reviews! I really appreciate the time you took to help with this.

@Yurlungur — I've opened a PR with some minor edits to the paper. After merging or responding to that, here are the final steps that I'll need from you:
