### 2022-07-05 6ae54a8
fix minor typos from final readthrough
