### 2022-07-05 #4367 @Yurlungur
Thanks @jzrake @lgarrison ! I appreciate the time you spent engaging with me and the project.

Thanks @dfm I've merged the PR, done a final readthrough, and incremented version number.
- Version number is: v1.5.1
- DOI is: 10.5281/zenodo.6800124
