### 2022-07-05 #4367 @dfm
@Yurlungur — I've now handed this off to the managing editors to do the final processing. There may be some final edits or other changes, but the process should be fairly quick. Thanks again for your submission and for your responses to all the suggestions from @lgarrison and @jzrake!

@lgarrison, @jzrake — Thanks again for your reviews of this submission. Thanks for the time that you took and the thorough and constructive comments that you made. We couldn't do this without you, and I really appreciate you volunteering your time!!

@openjournals/joss-eics — Some notes: (1) @jzrake's checklist does not have the performance box checked because he didn't have the required hardware (see comment: https://github.com/openjournals/joss-reviews/issues/4367#issuecomment-1170046215), but @lgarrison was able to replicate the results so I'm happy with this as it stands; (2) the bot has gone rogue with it's recommended missing DOI above... I'm not sure what it's thinking!
