+++ paper.bib
@@ -1,6 +1,6 @@
 @ARTICLE{fornax,
        author = {{Skinner}, M. Aaron and {Dolence}, Joshua C. and {Burrows}, Adam and {Radice}, David and {Vartanyan}, David},
-        title = "{FORNAX: A Flexible Code for Multiphysics Astrophysical Simulations}",
+        title = {{FORNAX}: A Flexible Code for Multiphysics Astrophysical Simulations},
       journal = {The Astrophysical Journal Supplement Series},
      keywords = {methods: numerical, Astrophysics - Instrumentation and Methods for Astrophysics, Astrophysics - High Energy Astrophysical Phenomena, Astrophysics - Solar and Stellar Astrophysics, Physics - Computational Physics},
          year = 2019,
@@ -38,7 +38,7 @@ archivePrefix = {arXiv},
 
 @techreport{sesame,
 author = "Lyon, Stanford P. and Johnson, James D.",
-title = "Sesame: The Los Alamos National Laboratory Equation of State Database",
+title = {Sesame: The {L}os {A}lamos {N}ational {L}aboratory Equation of State Database},
 institution = "Los Alamos National Laboratory",
 number="LA-UR-92-3407",
 year = "1992"}
@@ -110,14 +110,14 @@ url = {https://github.com/catchorg/Catch2},
 
 @Misc{HDF5,
     author = {{The HDF Group}},
-    title = "{Hierarchical data format version 5}",
+    title = {Hierarchical data format version 5},
     year = {2000},
     url = {http://www.hdfgroup.org/HDF5}
 }
 
 @ARTICLE{SullivanWeak,
        author = {{Sullivan}, Chris and {O'Connor}, Evan and {Zegers}, Remco G.~T. and {Grubb}, Thomas and {Austin}, Sam M.},
-        title = "{The Sensitivity of Core-collapse Supernovae to Nuclear Electron Capture}",
+        title = {The Sensitivity of Core-collapse Supernovae to Nuclear Electron Capture},
       journal = {The Astrophysical Journal},
      keywords = {neutrinos, nuclear reactions, nucleosynthesis, abundances, equation of state, supernovae: general, Astrophysics - High Energy Astrophysical Phenomena, Nuclear Experiment, Nuclear Theory},
          year = 2016,
@@ -166,5 +166,5 @@ archivePrefix = {arXiv},
   title={Parallel programming in OpenMP},
   author={Chandra, Rohit and Dagum, Leo and Kohr, David and Menon, Ramesh and Maydan, Dror and McDonald, Jeff},
   year={2001},
-  publisher={Morgan kaufmann}
+  publisher={Morgan Kaufmann}
 }
