### 2022-07-05 de33e32
3 more small changes
sorry I missed these the first time - this should be it
